plugin restructure for movies and more
