+++ internal/config/config.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/mitchellh/mapstructure"
 	"github.com/rbtr/pachinko/internal/pipeline"
+	internalpre "github.com/rbtr/pachinko/internal/plugin/processor/pre"
 	"github.com/rbtr/pachinko/plugin/input"
 	"github.com/rbtr/pachinko/plugin/output"
 	"github.com/rbtr/pachinko/plugin/processor"
@@ -27,18 +28,18 @@ type RootConfig struct {
 
 type ConfigCmdConfig struct {
 	RootConfig `mapstructure:",squash"`
-	Format     string   `mapstructure:"format"`
-	Inputs     []string `mapstructure:"inputs"`
-	Outputs    []string `mapstructure:"outputs"`
-	Processors []string `mapstructure:"processors"`
+	Format     string                      `mapstructure:"format"`
+	Inputs     []string                    `mapstructure:"inputs"`
+	Outputs    []string                    `mapstructure:"outputs"`
+	Processors map[processor.Type][]string `mapstructure:"processors"`
 }
 
 type SortCmdConfig struct {
 	RootConfig `mapstructure:",squash"`
-	Pipeline   pipeline.Config          `mapstructure:"pipeline"`
-	Inputs     []map[string]interface{} `mapstructure:"inputs"`
-	Outputs    []map[string]interface{} `mapstructure:"outputs"`
-	Processors []map[string]interface{} `mapstructure:"processors"`
+	Pipeline   pipeline.Config                             `mapstructure:"pipeline"`
+	Inputs     []map[string]interface{}                    `mapstructure:"inputs"`
+	Outputs    []map[string]interface{}                    `mapstructure:"outputs"`
+	Processors map[processor.Type][]map[string]interface{} `mapstructure:"processors"`
 }
 
 func (c *RootConfig) configLogger() {
@@ -102,17 +103,21 @@ func (c *SortCmdConfig) ConfigurePipeline(pipe *pipeline.Pipeline) error {
 		}
 	}
 
-	for _, p := range c.Processors {
-		if name, ok := p["name"]; ok {
-			if initializer, ok := processor.Registry[name.(string)]; ok {
-				plugin := initializer()
-				if err := mapstructure.Decode(p, plugin); err != nil {
-					return err
+	pipe.WithProcessors(internalpre.NewCategorizer())
+
+	for _, t := range processor.Types {
+		for _, p := range c.Processors[t] {
+			if name, ok := p["name"]; ok {
+				if initializer, ok := processor.Registry[t][name.(string)]; ok {
+					plugin := initializer()
+					if err := mapstructure.Decode(p, plugin); err != nil {
+						return err
+					}
+					if err := plugin.Init(); err != nil {
+						return err
+					}
+					pipe.WithProcessors(plugin)
 				}
-				if err := plugin.Init(); err != nil {
-					return err
-				}
-				pipe.WithProcessors(plugin)
 			}
 		}
 	}
@@ -120,9 +125,19 @@ func (c *SortCmdConfig) ConfigurePipeline(pipe *pipeline.Pipeline) error {
 	return nil
 }
 
+func NewSortCmdConfig() *SortCmdConfig {
+	return &SortCmdConfig{
+		Processors: map[processor.Type][]map[string]interface{}{
+			processor.Pre:   {},
+			processor.Intra: {},
+			processor.Post:  {},
+		},
+	}
+}
+
 // LoadConfig loadconfig
 func LoadSortCmdConfig() (*SortCmdConfig, error) {
-	cfg := &SortCmdConfig{}
+	cfg := NewSortCmdConfig()
 	viper.SetEnvKeyReplacer(strings.NewReplacer("_", "-"))
 	viper.AutomaticEnv()
 	err := viper.Unmarshal(cfg)
@@ -138,8 +153,10 @@ func (c *ConfigCmdConfig) DefaultConfig(p *SortCmdConfig) error {
 		for k := range output.Registry {
 			c.Outputs = append(c.Outputs, k)
 		}
-		for k := range processor.Registry {
-			c.Processors = append(c.Processors, k)
+		for _, t := range []processor.Type{processor.Pre, processor.Post, processor.Intra} {
+			for k := range processor.Registry[t] {
+				c.Processors[t] = append(c.Processors[t], k)
+			}
 		}
 	}
 
@@ -167,15 +184,17 @@ func (c *ConfigCmdConfig) DefaultConfig(p *SortCmdConfig) error {
 		}
 	}
 
-	for _, name := range c.Processors {
-		log.Tracef("making default config for plugin %s", name)
-		if initializer, ok := processor.Registry[name]; ok {
-			var out map[string]interface{}
-			if err := mapstructure.Decode(initializer(), &out); err != nil {
-				return err
+	for t, names := range c.Processors {
+		for _, name := range names {
+			log.Tracef("making default config for plugin %s", name)
+			if initializer, ok := processor.Registry[t][name]; ok {
+				var out map[string]interface{}
+				if err := mapstructure.Decode(initializer(), &out); err != nil {
+					return err
+				}
+				out["name"] = name
+				p.Processors[t] = append(p.Processors[t], out)
 			}
-			out["name"] = name
-			p.Processors = append(p.Processors, out)
 		}
 	}
