+++ internal/plugin/processor/pre/types.go
@@ -1,0 +1,15 @@
+/*
+Copyright © 2020 The Pachinko Authors
+
+This Source Code Form is subject to the terms of the Mozilla Public
+License, v. 2.0. If a copy of the MPL was not distributed with this
+file, You can obtain one at https://mozilla.org/MPL/2.0/.
+*/
+
+/*
+Internal package processor (pre/post/intra) provides implementations of
+unconfigurable, non-optional processor plugins that are tied to the run
+of certain public plugins but are not specific enough to be integrated in to
+those plugins.
+*/
+package pre
