+++ plugin/include.go
@@ -8,12 +8,9 @@ file, You can obtain one at https://mozilla.org/MPL/2.0/.
 package plugin
 
 import (
-	_ "github.com/rbtr/pachinko/plugin/input/path"
-	_ "github.com/rbtr/pachinko/plugin/output/move"
-	_ "github.com/rbtr/pachinko/plugin/output/stdout"
-	_ "github.com/rbtr/pachinko/plugin/processor/tvdb"
-	_ "github.com/rbtr/pachinko/plugin/processor/tvdest"
-	_ "github.com/rbtr/pachinko/plugin/processor/tvid"
-	_ "github.com/rbtr/pachinko/plugin/processor/tvmeta"
-	_ "github.com/rbtr/pachinko/plugin/processor/vidid"
+	_ "github.com/rbtr/pachinko/plugin/input"
+	_ "github.com/rbtr/pachinko/plugin/output"
+	_ "github.com/rbtr/pachinko/plugin/processor/intra"
+	_ "github.com/rbtr/pachinko/plugin/processor/post"
+	_ "github.com/rbtr/pachinko/plugin/processor/pre"
 )
